+++ quiz-service/src/main/java/com/example/quiz_service/repository/QuizRepository.java
@@ -3,5 +3,5 @@ package com.example.quiz_service.repository;
 import com.example.quiz_service.model.Quiz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuizRepository extends JpaRepository<Quiz, Integer> {
+public interface QuizRepository extends JpaRepository<Quiz, Long> {
 }
